Don't use the implicit jdbc connection during councurrency tests.
Use a worker thread to insert bootstrap data so that we are sure the
content is flushed and available in the database for other workers
to find.

The problem with using the main thread is that data inserted in it
(in all Hibernate Sessions that share the associated JDBC connection) is
not committed to the database. This is because Grails needs to roll back
data between tests.